+++ CHANGELOG.md
@@ -13,7 +13,9 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
  - README and examples/README to cover local builds, setup and test execution
  - Go verison to 1.14 and against Go master branch in Travis CI
  - Travis CI file execute to execute website-test covering provider documentation
+ - Provider uses Terraform Plugin SDK V1
  
 ### Removed
  - Removed storage HCL blocks that failed due to referencing older UpCloud template ID
+ - Removed the plan, price, price_zone and timezone UpCloud resources